Move math methods for availability into separated class

+++ framework/src/main/java/com/bytex/snamp/connector/metrics/Arrivals.java
@@ -11,7 +11,7 @@ public interface Arrivals extends RatedTimer {
 
     double getInstantAvailability();
 
-    long getChannels();
+    int getChannels();
 
     /**
      * Gets ratio between summary duration of all requests and server uptime.
